requirements.txt
backports.tarfile==1.2.0
bcrypt==4.2.1
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
croniter==6.0.0
cryptography==44.0.0
cycler==0.12.1
et_xmlfile==2.0.0
Flask==3.1.0
fonttools==4.55.3
func-timeout==4.3.5
future==1.0.0
hnmp==0.5.1
idna==3.10
igraph==0.11.8
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.4
junos-eznc==2.7.2
keyring==25.5.0
kiwisolver==1.4.7
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
mdurl==0.1.2
more-itertools==10.5.0
n2g==0.3.3
napalm==5.0.0
napalm-arubaos-switch==0.2.0
napalm-procurve==0.6.0
ncclient==0.6.15
netaddr==1.3.0
netmiko==4.5.0
netutils==1.10.0
networkx==3.2.1
ntc_templates==7.5.0
numpy==2.0.2
openpyxl==3.1.5
packaging==24.2
paramiko==3.5.0
pillow==11.0.0
ply==3.11
pyaoscx==2.6.0
pyasn1==0.4.8
pycparser==2.22
pycryptodomex==3.21.0
pyeapi==1.0.4
Pygments==2.18.0
PyNaCl==1.5.0
pynxos==0.0.5
pyparsing==3.2.0
PyQt6==6.8.0
PyQt6-Charts==6.8.0
PyQt6-Charts-Qt6==6.8.1
PyQt6-Qt6==6.8.1
PyQt6-WebEngine==6.8.0
PyQt6-WebEngine-Qt6==6.8.1
PyQt6_sip==13.9.1
pyserial==3.5
pysmi==1.5.9
pysnmp==4.4.12
python-dateutil==2.9.0.post0
python-igraph==0.11.8
pytz==2024.2
pywin32-ctypes==0.2.3
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
scp==0.15.0
six==1.17.0
textfsm==1.1.3
texttable==1.7.0
transitions==0.9.2
ttp==0.9.5
ttp-templates==0.3.7
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
yamlordereddictloader==0.4.2
zipp==3.21.0
